software/src/requirements.txt
pyaml
termcolor
matplotlib
numpy
scipy
